# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/purchase_agreements:
    get:
      tags:
        - PurchaseAgreements
      summary: List PurchaseAgreements
      description: Get a paginated list of PurchaseAgreements with optional filtering and sorting
      operationId: listPurchaseAgreements
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseAgreementList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseAgreements
      summary: Create PurchaseAgreement
      description: Create a new PurchaseAgreement
      operationId: createPurchaseAgreement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseAgreementRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseAgreement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/purchase_agreement_lines:
    get:
      tags:
        - PurchaseAgreementLines
      summary: List PurchaseAgreementLines
      description: Get a paginated list of PurchaseAgreementLines with optional filtering and sorting
      operationId: listPurchaseAgreementLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseAgreementLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseAgreementLines
      summary: Create PurchaseAgreementLine
      description: Create a new PurchaseAgreementLine
      operationId: createPurchaseAgreementLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseAgreementLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseAgreementLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PurchaseAgreement:
      type: object
      required:
        - id
        - agreement_number
        - agreement_kind
        - status
        - company_id
        - supplier_id
        - currency
        - metadata
      properties:
        id:
          type: string
          format: uuid
        agreement_number:
          type: string
          maxLength: 40
        agreement_kind:
          type: string
        status:
          type: string
        company_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        date_start:
          nullable: true
          type: string
          format: date
        date_end:
          nullable: true
          type: string
          format: date
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePurchaseAgreementRequest:
      type: object
      required:
        - agreement_number
        - agreement_kind
        - status
        - company_id
        - supplier_id
        - currency
        - metadata
      properties:
        agreement_number:
          type: string
          maxLength: 40
        agreement_kind:
          type: string
        status:
          type: string
        company_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        date_start:
          nullable: true
          type: string
          format: date
        date_end:
          nullable: true
          type: string
          format: date
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PurchaseAgreementList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseAgreement'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PurchaseAgreementLine:
      type: object
      required:
        - id
        - agreement_id
        - company_id
        - item_id
        - quantity
        - rate
        - qty_ordered
        - metadata
      properties:
        id:
          type: string
          format: uuid
        agreement_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        rate:
          type: string
        qty_ordered:
          type: string
        metadata:
          type: object
...
    CreatePurchaseAgreementLineRequest:
      type: object
      required:
        - agreement_id
        - company_id
        - item_id
        - quantity
        - rate
        - qty_ordered
        - metadata
      properties:
        agreement_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        rate:
          type: string
        qty_ordered:
          type: string
        metadata:
          type: object
...
    PurchaseAgreementLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseAgreementLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []